requirements.txt
Flask==0.12.0
Jinja2==2.8.1
gunicorn==19.6.0
python-telegram-bot==9.0.0
